Feature/APMI3255 iOS: options factory update
APMI-3255:-
1. Set Location method that sets global attributes (only 2 numbers - latitude and longitude)
2. Bundle identifier (CF Bundle) that is detected and added automatically.
3. Feature flag for slow/frozen renders
4. Contents attribute added.
